replace inotify by shell script
this change updates the minimum required go lang version in one of the
developer docs, and also changes the method for downloading the source
code. it should produce fewer warnings for developers wishing to install
operator-sdk from source. it also adds a note about needing GOPROXY set
to its default.